feat: support putRecords
