+++ src/index.js
@@ -1,5 +1,6 @@
 const arnParser = require('./arnParser');
 const emitEvent = require('./event');
+const emitEvents = require('./events');
 const validate = require('./validate');
 const instantiateKinesis = require('./kinesis');
 
@@ -16,6 +17,20 @@ module.exports = (config = {}) => {
   if (config.endpoint) {
     kinesis.endpoint = config.endpoint;
   }
+
+  if (config.type && config.type === 'BATCH') {
+    return (events) => {
+
+      if (!Array.isArray(events)) throw 'Events needs to be an Array.';
+
+      if (!events.length) throw 'Events are missing.';
+
+      // Each PutRecords request can support up to only 500 records.
+      if (events.length > 500) throw 'Events array can only have 500 records';
+  
+      return emitEvents(kinesis, events, extendedConfig);
+    };
+  } 
 
   return (event) => {
     if (!event) throw 'Event is missing.';
